🎨 better end output structure
